+++ flutter_modular/lib/src/presenters/navigation/modular_router_delegate.dart
@@ -131,6 +131,7 @@ class ModularRouterDelegate extends RouterDelegate<ModularRoute>
           }
 
           _pages.remove(_lastPageModule);
+          _pages.add(page);
         }
       }
     }